implicit: Split decl/impl.
